[Refactor] Seperate OpOverrides
